added "test-history.sh" to check that the temporary history mechanism works
It doesn't work in old versions of bash because
    $ fc -l
ignores the last command, even if it is not of the form "fc ..."

Since we have a history where all commands where inserted manually, when
running
    $ fc -l
the last command is NOT "fc -l", so that we are missing a command!

This happens in macOS' bash (version 3.2).

This shouldn't be problematic as this mechanism is only used when
running "auto.sh" or "test.sh" scripts.

also, erase history when creating a temporary history in tests scripts
in bash. (This is automatic in zsh.)